Add deleteApps function
The deleteApps function is invoked when the user selects the destroy command from the command line. It queries the DynamoDB for information on the Preview Apps which are active, retrieves the necessary GitHub owner, repo, and PAT info, and then uses a POST request to trigger the GitHub action that tears down all active preview apps.